#155 process is undefined in webpack
process is a global nodejs variable that is not defined in the browser by default. So webpack doesn't know about it. Of course there are ways to define the process variable with webpack but that is not the default behavior and needs a plugin (DefinePlugin) or other workaround.
I've tested this change locally with an angular/ionic app and it works great.